replace placeholder and reorder files
