+++ src/app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/UploadPhotosForm.tsx
@@ -4,13 +4,13 @@ import { GlobalStore } from "app/store/GlobalStore"
 import { showPhotoActionSheet } from "app/utils/requestPhotos"
 import { useFormikContext } from "formik"
 import { Button, Flex, Spacer, Text } from "palette"
+import { PhotoRow } from "palette/elements/PhotoRow/PhotoRow"
+import { Photo, PhotosFormModel } from "palette/elements/PhotoRow/validation"
 import React from "react"
 import { removeAssetFromSubmission } from "../Mutations/removeAssetFromConsignmentSubmissionMutation"
-import { PhotoRow } from "./PhotoRow"
 import { addPhotoToConsignment } from "./utils/addPhotoToConsignment"
 import { calculateSinglePhotoSize } from "./utils/calculatePhotoSize"
 import { isSizeLimitExceeded } from "./utils/calculatePhotoSize"
-import { Photo, PhotosFormModel } from "./validation"
 
 export const UploadPhotosForm: React.FC<{ isAnyPhotoLoading?: boolean }> = ({
   isAnyPhotoLoading,
