+++ src/palette/elements/PhotoRow/PhotoRow.tests.tsx
