+++ src/palette/elements/PhotoRow/validation.ts
